requirements.txt
discord.py
googletrans
PyNaCl
python-dotenv
Requests
youtube_dl
deep-translator
langdetect
langcodes
pymongo[srv]
certifi